// frontend/src/types/api-schema.d.ts
export interface components {
    schemas: {
// ...
        PageResponseDto: {
            /**
             * @description Identificador único da página
             * @example 123e4567-e89b-12d3-a456-426614174000
             */
            id: string;
            /**
             * @description Nome para exibição da página
             * @example Galaxy Bay Official
             */
            name: string;
            /**
             * @description Slug único para URL
             * @example galaxy-bay-official
             */
            slug: string;
            /**
             * @description Tagline curta da página
             * @example O melhor lugar para explorar o universo NFT.
             */
            tagline?: string;
            /**
             * @description Descrição completa em texto rico
             * @example Aqui você encontra …
             */
            description?: string;
            /**
             * @description URL do avatar da página
             * @example https://cdn.exemplo.com/avatar.png
             */
            avatarUrl?: string;
            /**
             * @description URL do banner da página
             * @example https://cdn.exemplo.com/banner.jpg
             */
            bannerUrl?: string;
            /**
             * @description Status da página (por exemplo: 'draft', 'published')
             * @example draft
             */
            status?: string;
            /**
             * @description Categoria da página
             * @example arte digital
             */
            category?: string;
            /**
             * @description Tags associadas à página
             * @example [
             *       "nft",
             *       "arte",
             *       "galáxia"
             *     ]
             */
            tags?: string[];
            /**
             * @description Total de visualizações da página
             * @example 123
             */
            viewCount: number;
            /**
             * @description Total de inscritos na página
             * @example 45
             */
            salesCount: number;
            /**
             * Format: date-time
             * @description Data de criação da página
             * @example 2025-05-11T14:23:00.000Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @description Data da última atualização da página
             * @example 2025-05-12T08:30:00.000Z
             */
            updatedAt: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}